+++ Changelog.md
@@ -1,1 +1,8 @@
+# v0.1.1, 2021-03
+
+Rust:
+- Configurable field order (`zki_sieve example --field-order=101`)
+- Example with an incorrect witness (`zki_sieve example --incorrect`)
+- Rename library to zki_sieve
+
 # v0.1.0, 2020-09